Isolate DSpark capacity scheduler metadata
